requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
Django==3.1.1
django-bootstrap5==2.0.0
django-crispy-forms==1.11.2
django-debug-toolbar==3.2.1
django-environ==0.4.5
django-filter==2.4.0
django-phonenumber-field==5.1.0
djangorestframework==3.12.4
idna==2.10
mysqlclient==2.0.3
phonenumbers==8.12.22
Pillow==8.2.0
pytz==2021.1
requests==2.25.1
soupsieve==2.2.1
sqlparse==0.4.1
urllib3==1.26.4
